atualizando

+++ 42cursus/push_swap/push_swap.h
@@ -19,12 +19,12 @@
 # include <stdbool.h>
 # include <stdlib.h>
 
-typedef struct s_stack
+typedef struct	s_stack
 {
-	long			nbr;
-	long			index;
-	struct s_stack	*next;
-	struct s_stack *prev;
+	int		nbr;
+	int		index;
+	struct	s_stack	*next;
+	struct	s_stack *prev;
 }	t_stack;
 
 #endif
